removed workaround for playing some older tapes for MM engine
This removes a workaround that added compatibility code for replaying
old tapes which used the game engine's random number generator for
drawing game elements, which generally should not happen.

This affects the tapes for levels 033 and 048 of "classic_mindbender",
which have been re-recorded with the above change.

This change reverts commit 8b514628.