added looking for game music directory by checking directory files
Before, the first matching music directory was used -- regardless of
its contents. This means that it may indeed contain game music files
(that are not configured for any other purpose), or that it is a level
specific music directory that only contains some additional music
configured for a certain purpose, but no game music files, in which
case no game music files were found and used at all (as the general
custom music directory with all the game music would not be used in
this case).

This change adds looking for the first matching music directory that
really contains game music (not configured for other purposes).