fixed bug that could cause an endless loop (freeze) in the MM engine
This is a workaround for a regression bug of commit f8aec447 (that
fixed overloading the laser due to duplicated tiles in damage list).

The problem that could cause an endless loop already existed before
the above commit, but without that commit, scanning the laser was
stopped by the damage list completely filling up, stopping the loop.
The above commit prevents adding the same tile again and again to the
damage list, therefore not stopping the endless loop anymore.

This commit should prevent the endless loop from happening at all.

The problem (and how to reproduce it) was described by Eizzoux in the
following post in the R'n'D web forum:
https://www.artsoft.org/forum/viewtopic.php?p=18214