improved handling laser on two diagonally adjacent elements even more
This also correctly handles the case of the laser hitting the inner
corner of two diagonally adjacent game elements with a third game
element behind these two elements (in the direction of the laser).

It also makes sure that no "inner corners" of diagonally adjacent
mini tiles (MM wall tiles) are handled, which are not interesting
here (as they are always the same type of wall), but only diagonal
corners between full tile positions.

This change also makes sure that randomly selecting which element to
handle is only done if really necessary (to not break existing tapes
by additional calls to the random number generator).