fixed rotating MM game elements in editor with pressed Shift key
When holding the Shift key pressed, MM style rotatable game elements
can be rotated on the playfield (instead of drawing with the element
selected for this mouse button). However, if the drawing element is
an MM style wall, this element is placed on the playfield instead of
rotating the existing element on the playfield.

This usually happened when editing a level using the MM game engine
and trying to rotate an existing element to the right using the right
mouse button, as the default drawing element for the right mouse
button is the wooden MM style wall.

This change fixes this problem, so all three mouse button can now
be used to rotate elements on the playfield, regardless of current
drawing element.