fixed reading music from directory if music configured for all levels
When collecting music files that may have music info available for the
music info screen, do not add music files from music directory if game
music is configured for all levels of a level set.

This fixes a bug with showing music info for music files that is never
played in the game, which happens if there is no "unconfigured" music
in a custom music directory (because all contained music is configured
in the custom music configuration), so the program falls back to the
default music directory (which contains unconfigured music files which
can be played as in-game music, but which is never used in this case).