fixed bug with scaled down background bitmaps
This change makes sure to always use the original sized bitmaps for
backgrounds, instead of the "standard size" bitmap size, which can
indeed be scaled down if the same image file was used not only for
a background bitmap, but also for a game element, but with a tile
size larger than 32x32 (which results in the image being scaled down
to also get a 32x32 tile size). When using the same image for
background graphics, the unmodified original bitmap size must be used
instead of the "normalized" or "standard" (scaled down) bitmap size.