fixed bug with playing two global game sounds as game element sounds
The two game sounds "game.losing" and "game.sokoban_solving" were
played as playfield position related, game element specific sounds
before (which especially caused problems with "game.losing" being
defined as a loop sound, but which was stopped after a very short
time, because loop sounds have to be regularly re-initiated to keep
playing, which was not the case here).

As a result, the sound effect "game.losing" (an evil laughter after
the player has died) could effectively not be heard anymore after a
game was lost for many years.