added respecting user-configured screen orientation locking on Android
On Android, locking the screen orientation to portrait or landscape
as configured by the user of the device was completely ignored so far.

This change adds respecting this device setting by using the setting
"fullUser" for the app's screen orientation activity (always use user
defined screen orientation) instead of the previously used setting
"fullSensor" (always use sensor-based rotation, regardless of any
user-configured locking of screen rotation).

The corresponding entry in the Android Manifest file was added only to
document this changed behaviour, even though it is in fact ignored, as
this activity setting is finally hard-coded in the Java activity code,
where it was changed to use the new behaviour.