fixed several bugs with shared playfield bitmap in MM game engine
This fixes several bugs that resulted from using the normal background
bitmap for the Mirror Magic game engine (which was possible because
this game engine does not support scrolling levels and therefore only
uses the visible playfield viewport).

The main bugs fixed by this change are preventing MM engine levels
that are larger than the visible playfield area drawing over the
border area, and some problems with envelope style requests (like
restoring an outdated copy of the backbuffer after an envelope
request was answered after a game was lost).

While fixing these visible bugs, this change also brings a cleanup of
the MM engine code, which used three different bitmap pointers before
(that all pointed to the same backbuffer bitmap).