fixed graphical bugs with off-screen playfield movement in MM engine
This changes handles graphical bugs that may occur if the playfield of
levels played with the MM game engine are larger than the visible part
of the playfield. (This is not supported with MM style levels, but may
occur when using custom artwork with smaller playfield viewport than
intended by the level designer.)

The problem could be seen with levels where pacman game elements leave
and re-enter the visible part of the playfield.