fixed regression bug with bad playfield offset
This reverts commit a35cd127.

The bug introduced by the above commit was caused by a wrong comment
indicating that a certain offset was related to the playfield border,
which was wrong. In fact, this offset is related to the width of the
laser beam (to store the middle position of the tile the laser beam
touches, to be able to draw a variable-width laser beam later with a
size that is relative to the middle of the tile).

As the laser beam currently has a width of 4 pixels, this value has
to be set to 2 here.