added playing the first unconfigured music file in the first level
Before, the (unconfigured) music file to be played in the first level
of a level set depended on the number of the first level (with the
result that the first music file was played in level 000, the second
music file was played in level 001 and so on, which probably was not
the intended behaviour when the level set started with level 001).

Now the first music file in the list of unconfigured music files is
played in the first level of a level set, regardless of the number of
that level (which is either 000 or 001 for most level sets).