fixed one-frame delay when global animations leave "init" state
This fixes a bug for global animations which either have an initial
delay or an initial event defined, causing a delay of one screen frame
after the initial waiting state has ended and before the running state
(which displays the global animation) is started.

This can especially be visible if a changing CE is graphically replaced
with a global animation triggered by the CE change.