improved forced restart of global animations when restarting game
When forcing a restart of global animations when restarting the game,
a full screen redraw is neither required nor desired. Therefore, this
change only restarts global animations according to the game status
passed to function "RestartGlobalAnimsByStatus()", but does not redraw
the screen (including global animations, so nothing is drawn at all).