fixed CE condition that checks touching the player at some side
In the level editor on the "change" tab of a custom element, when
using the CE condition "[x] element changes ..." in combination with
"[x][touching] element [<element>]" with "at [<side>] side", there
was wrong behaviour in some cases when "<element>" was the player.

This was caused by not checking the side of the player (that was being
touched by the custom element), but by checking the side of the custom
element (touching the player), which is the exact opposite side here.

This change fixes the bug, now always checking the correct side as it
is configured in the level editor. (Regarding to the game engine test
suite, this functional game engine change did not break any tapes.)