added new animation mode "level_nr" to select frame according to level
This change adds a new animation mode "level_nr" which selects a
single animation frame from a number of frames that corresponds to the
current level. That is, the first frame is used for the first level,
the second frame is used for the second level and so on (overwrapping
frames if there are more levels than animation frames). This animation
mode can therefore only be used for non-animated game graphics, like
walls. So this animation mode can be used to have different graphics
in different levels for certain game elements, while still using the
same game elements in each level.