added selecting element for group elements according to current level
This change works similar to the previous commit, but not selecting an
animation frame according to the current level, but selecting a game
element from a number of elements in a group element (selecting the
first element for the first level, the second element for the second
level, and so on, again overwrapping if there are more levels than
game elements defined for the group element). That way, different game
elements can be used in different levels which all use the same group
elements.