fixed bug with CE change setting the player to its current position
This change fixes a bug where a CE change places the player at the
current player position, causing an explosion at that playfield
position, which should not happen.

In other words, such an event would normally relocate the player, but
to a playfield position which is identical to the current playfield
position of the player, which should therefore have no effect at all.

The bug occured because it was not checked if the new element to be
created is the player itself, but only checks if an element that is
not walkable by the player is created at the players current position.