fixed half-tile shifting when teleporting between playfield borders
When relocating the player without centering the screen and without
scrolling ("teleporting" to a different playfield area, but keeping
the screen position), shifting the visible part of the playfield by
half a tile may happen if the screen width or height is even.

This is especially true when teleporting from and to an area that is
directly aligned with the border of the playfield (regardless of the
tiles at the playfield border being only half or fully visible, which
only happens with even sized screen width or height, depending on the
current player scroll position).

This change makes sure that both the source and destination area of
the playfield are displayed either aligned by half a tile or a full
tile when doing a "teleporting" player relocation in such a case.