fixed bug with loading artwork if only one special artwork set exists
R'n'D supports using additional special artwork, like "ECS" and "AGA"
graphics sets or "default" and "lowpass" sound sets, which can be
selected from the game engines setup menu. This works fine if both
artwork sets are defined (and also if no special artwork is defined).

However, it did not work correctly if only one special artwork set was
defined, in addition to the "standard" artwork set, like this:

graphics_set:                   gfx_bond_mine_01
graphics_set.aga:               gfx_bond_mine_01.aga

In this case, it was not possible to use the "standard" graphics set,
but only the "AGA" graphics set (because the "standard" graphics set
was internally overridden by the "AGA" graphics set when selecting to
use "AGA" graphics in the game engines setup menu, and selecting to
use "ECS" graphics later could not recover this). (This bug could be
seen in level set "emc_bond_mine_01" when starting the game with "AGA"
graphics preferred in the setup menu, which resulted in not being able
to select the "standard" graphics by setting graphics preference to
"ECS" in the game engines setup menu.)

This is fixed now (by internally setting the "ECS" artwork set, which
is not explicitly defined, to the "standard" artwork before internally
overriding it with either the "ECS" or "AGA" graphics set).

This bug was caused by commit 4bc6719c.