added random push delay for rocks, nuts and bombs to EM game engine
Just like in the original Emerald Mine game, there is now a random
delay when trying to push a rock, a nut or a bomb. While the chance
for rocks being successfully pushed at first attempt will be around
50 %, the chance for nuts will be a bit lower than that for rocks,
and the chance for bombs will be a bit lower than that for nuts,
with a roughly similar overall delay as in the original game.