significant cleanup of envelope style request dialog handling
After this change, the request dialog is not drawn to the back buffer
of the screen anymore, but to a temporary bitmap buffer (with the
request buttons directly being updated there after button events),
which is now only drawn during the final composition of the screen.

Therefore, the back buffer is not touched anymore when handling the
request dialog, and also does not have to be restored afterwards.