added forcing user to restart program after adding level collection
When adding a level group or an artwork group by dropping a zip file
into the program window, show a request dialog that the program has to
be restarted to activate the new content.

Unfortunately, this is currently required to rebuild the internal
level and artwork set structures (and should be improved to be rebuild
them on-the-fly without a program restart).

Before, the program was just stopped with a user-unfriendly message
reporting a fatal error.