changed order of processing game modes for global animations clicks
The previous (reversed) order suggested that this had something to do
with draw order, which was not the case. Using the normal (numerical)
order also has the advantage of processing global animations without
a special game mode assigned to them (like ".PLAYING" or ".MAIN") to
be processed first, which is especially important to process pointer
style global animations (which are probably defined to be used on all
game screens in most cases) before all other global animations, which
is required to display an alternative pointer when the mouse button
is pressed (which is defined using ".style: passthrough_clicks" for
such cases).