reverted order of processing game modes for global animations clicks
The assumptions that led to the change that is reverted here were
simply wrong: When drawing global animations in a certain order,
checking for mouse clicks must be done in reverse order, to ensure
that the click is processed for the topmost global animation. This
includes processing game modes in reverse order (which are handled
in normal order when drawing global animations).

Processing clicks for a global animation based custom mouse pointer
should be handled separately anyway, to process it independently
from the game mode it was defined for.

This reverts commit e3411dc9.