re-added explicitly setting draw order for pointer-style animations
Even though pointer-style animations are handled specially when being
drawn (on top of all other global animations), they are not handled
specially when being checked for clicks. Therefore, forcing a high
draw order for pointer-style animations is still required (as the draw
order is not only used for drawing, but also for the order of checking
animations for mouse clicks).

This reverts commit d1a0dcb3.