added scrolling without iterating BD engine if player out of sight
This is needed to prevent broken tapes in case of viewport or tile
size changes and may occur in case of wrap-around at level borders.

Before, recording tape events was continued during scrolling, even
though the BD game engine did not iterate while scrolling the cave
to the visual player position. This could result in broken tapes if
the tape was replayed with a different playing viewport or with a
different tile size (like using "small game elements" from the setup
menu), as this would have an effect on the duration of scrolling,
causing tape events and game engine running out of sync.