fixed bugs with wrong newline handling in envelopes (game and editor)
This is the second (and hopefully final) attempt to fix a bug that was
already thought to be fixed with commit e567cb7f (which only fixed the
bug partially, and introduced new problems).

See the detailed description in function "LoadLevel_FixEnvelopes()"
for more details on the problem and the solution.