fixed graphical bug with snapping player drawn as moving into explosion
This fixes a graphical bug with smooth movement animation in the BD
engine: When the player is snapping a tile that immediately after is
replaced by an explosion next to the player, the player was drawn as
moving into that explosion (without actually moving at all, as he was
only snapping). This change fixes this bug by checking if the player
actually moved.