fixed code for drawing diagonal player movement in BD engine
When drawing diagonal player movement, do not use smooth movement for
the player, and do not draw the source tile (but only draw the target
tile, as the player "jump-moves" to the next tile in this case).

So far, the source tile was not drawn at all due to not correctly
detecting diagonal movement, causing smooth movement being disabled
for this tile, so the player was not drawn at the source tile anyway
(which causes another bug yet to be fixed: not drawing tiles entering
the tile left by the player at the same time using smooth movement).

This change fixes (not) drawing the source tile for diagonal player
movement.