fixed pushing element from left side onto conveyor belt in BD engine
This prevents game elements that are pushed onto a conveyor belt from
the left side to move two tiles at once. This is a result of the cave
scan order (already pushed element moved another tile when conveyor
belt is handled during scan of the next cave row). Although unlikely,
this change affects the game engine and therefore may break existing
caves (and will break existing replays that use this case).

So far, this is the only change that affects the game engine that was
made to fix weird behavior when using smooth movement animations.