fixed graphical bug when pushing rocks onto BD style conveyor belts
When pushing a rock onto a right-moving conveyor belt from the left
side using the BD engine and using the "buggy conveyor belts" editor
setting (used by native BD caves by default), the player pushing that
rock was displayed with a broken animation when using "smooth movement
animations" for the BD engine.

This change fixes this graphical bug.

(The first part fixes placing a space, which is not a movement,
while the second part seems not to be required otherwise anyway.)