added setting increased BD cave time also in main game engine
This change is required if the BD cave time was increased by
collecting clocks; else the game is running out of time too early in
the main game engine, which still uses the previous level time, even
though the BD cave time is displayed correctly in the game panel.

However, a decreased BD cave time (for example, if a voodoo doll was
destroyed) may not be set in the main engine, as it may result in an
immediate "out of time" event, which is handled separately in the main
game engine.