fixed bug with sounds not being played in native BD sound engine
This fixes a bug where sounds (when using the native BD sound engine)
were not played because other sounds with higher precedence were
already playing. However, this should only be true for two sounds to
be played at the same time, while a sound with lower precedence may
be played if another sound with higher precedence is already playing
for some time.