added GIC flag to keep game panel door open when restarting game
Setting this flag to "true" in file "graphicsinfo.conf" prevents
closing the game panel door when starting a game while already being
in game mode "playing" (like restarting a level or playing the next
level when not showing the hall of fame screen between two levels).
This can be used to skip closing and opening the game panel door if
nothing happens between closing and opening (besides updating the
panel content), which can be the case with BD engine levels which use
screen covering and uncovering instead of fading out and fading in.