moved code to different function
The code to prevent instantly killing the player by falling/flying game
elements at wrap-around cave position was moved to a different function
of the native BD game engine.

Before, this check was done right before moving a game element from the
very right or bottom playfield position to the corresponding wrap-around
cave position (at the very left or top playfield position), instantly
killing the player in the next cave scan if directly under this element.

Now, this check is done at an earlier stage when checking if the falling
or flying game element can move to this playfield position at all (which
will be prevented if this playfield position contains empty space that
was occupied by a non-space game element before the current cave scan).