added element palette section for Krissz engine BD elements in editor
This game element palette section does not contain any new elements,
but just the subset of existing game elements of the native Boulder
Dash game engine also supported by the original Krissz engine.

So this addition is just for convenience when creating or editing
caves that would also be playable in the original Krissz engine.