fixed Krissz engine style amoeba conversion by magic wall in same frame
This change improves the Krissz engine style behavior of converting
the amoeba to diamonds during the same cave scan as activating the
magic wall, now also covering the case that the amoeba was able to
grow in the same cave scan (before the magic wall was activated).

This fixes commit 7fdfb46f.