fixed player leaving visible playfield when using infinite scrolling
The previously used solution totally ignored that the player may be
able to leave the visible part of the playfield when running faster
than the screen is scrolling, even when using infinite scrolling.

The current solution makes sure than crossing the playfield borders
does not count as "leaving the visible part of the playfield" when
using infinite scrolling.