fixed correcting scrollbar position if outside playfield viewport
This change prevents correcting the scrollbar position if it uses the
viewport border (that is not used by the playfield content). Usually,
the playfield border is very small, but can be defined to be larger
(especially for larger screens), so there may be enough room for the
scrollbar in the border area of the viewport.