fixed bug which causes wrong score and time after solving level
This is a terrible bug that only happens if the "frames" part of the
level time counter (counted in frames and seconds) overflows exactly
at the moment when the level is solved, resulting in a wrong score
and time (that is one second or 50 frames too low) and therefore in a
better ranking in the high score table.

As the game uses 50 frames per second, about 2 % of all existing score
entries are affected by this bug (which can be fixed for the scores on
the high score server, but not for the locally stored personal scores).