changed getting identcode for namespaced levels from level number
With this change, various functions do not use the current level's
identcode anymore, but the identcode of the level that was requested
by its level numner.

This change fixes the following problem: When showing the high scores
after solving a level, the game has already loaded the next level.
For namespaced levels, this caused loading the scores for the next
level (referenced by the already loaded identcode for the next level)
instead of showing the scores for the (previous) level just solved.

Now the scores are loaded for the identcode of the level as requested
by its level number.